requirements.txt
django
djangorestframework
markdown
django-filter
pandas
xlrd
hazm
pymongo[srv]
cleantext
clean-text
tqdm
scikit-learn
sentence-transformers
unidecode
